Use inttypes.h macros when formatting uint64_t, int64_t values
Conflicts:
	lib/sysfs.c
	misc-utils/lsblk.c
	sys-utils/swapon.c

+++ disk-utils/mkfs.minix.c
@@ -543,7 +543,7 @@ static void setup_tables(void) {
 	memset(inode_buffer,0, get_inode_buffer_size());
 	printf(_("%lu inodes\n"), inodes);
 	printf(_("%lu blocks\n"), zones);
-	printf(_("Firstdatazone=%jd (%jd)\n"), get_first_zone(), first_zone_data());
+	printf(_("Firstdatazone=%jd (%jd)\n"), (intmax_t)get_first_zone(), (intmax_t)first_zone_data());
 	printf(_("Zonesize=%zu\n"), (size_t) MINIX_BLOCK_SIZE << get_zone_size());
 	printf(_("Maxsize=%zu\n\n"),get_max_size());
 }
